+++ CHANGELOG.md
@@ -6,10 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ## [Unreleased]
+
 ### Added
+- Connection in session
 - Adding CHANGELOG.md
 - Moving to lein-git-version 1.2.7.
 
+### Changed
+- Refactor parsing/processing.
+- Move to instaparse for directive parsing.
 
 ## [0.0.12] -2017-01-27
 ### Added